+++ modules/dice/roll.js
@@ -2,6 +2,7 @@ import PopoutEditor from "../popout-editor.js";
 import { ForceDie } from "./dietype/ForceDie.js";
 import {migrateDataToSystem} from "../helpers/migration.js";
 import {ItemFFG} from "../items/item-ffg.js";
+import { applyMessageMode, getMessageMode } from "../helpers/chat.js";
 
 /**
  * New extension of the core DicePool class for evaluating rolls with the FFG DiceTerms
@@ -342,27 +343,28 @@ export class RollFFG extends Roll {
 
   /* -------------------------------------------- */
   /** @override */
-  async toMessage(messageData = {}, { rollMode = null, create = true } = {}) {
+  async toMessage(messageData = {}, { messageMode, rollMode, create = true } = {}) {
     // Perform the roll, if it has not yet been rolled
     if (!this._evaluated) await this.evaluate();
 
-    const rMode = rollMode || messageData.rollMode || game.settings.get("core", "rollMode");
-
-    if (["gmroll", "blindroll"].includes(rMode)) {
-      messageData.whisper = ChatMessage.getWhisperRecipients("GM");
-    }
-    if (rMode === "blindroll") messageData.blind = true;
-    if (rMode === "selfroll") messageData.whisper = [game.user.id];
+    const mode = getMessageMode({
+      messageMode: messageMode ?? rollMode ?? messageData.messageMode ?? messageData.rollMode,
+    });
 
     // Prepare chat data
     messageData = foundry.utils.mergeObject(
       {
-        user: game.user.id,
+        author: game.user.id,
         content: this.total,
         sound: CONFIG.sounds.dice,
       },
       messageData
     );
+    // Accept older macro data, but only pass current document fields to Foundry.
+    if (messageData.user) messageData.author = messageData.user;
+    delete messageData.user;
+    delete messageData.rollMode;
+    delete messageData.messageMode;
     messageData.rolls = [this];
 
     Hooks.call("ffgDiceMessage", this);
@@ -370,10 +372,11 @@ export class RollFFG extends Roll {
     // Either create the message or just return the chat data
     const cls = getDocumentClass("ChatMessage");
     const msg = new cls(messageData);
-    if (rMode) msg.applyRollMode(rMode);
+    if (mode) applyMessageMode(msg, mode);
 
     // Either create or return the data
-    return create ? await cls.create(msg) : msg;
+    const data = msg.toObject();
+    return create ? await cls.create(data) : data;
   }
 
   /** @override */