+++ modules/helpers/modifiers.js
@@ -1,3 +1,4 @@
+import EffectHelpers from "./effects.js";
 import PopoutModifiers from "../popout-modifiers.js";
 
 export default class ModifierHelpers {
@@ -734,7 +735,7 @@ export default class ModifierHelpers {
         for (const curMod of explodedMods) {
           changes.push({
             key: ModifierHelpers.getModKeyPath(curMod['modType'], curMod['mod']),
-            mode: CONST.ACTIVE_EFFECT_MODES.ADD,
+            ...EffectHelpers.changeType(),
             value: formData.data.attributes[k].value,
           });
         }