+++ modules/helpers/tours.js
@@ -1,3 +1,4 @@
+import EffectHelpers from "./effects.js";
 class CharacterTour extends foundry.nue.Tour {
   #originalDeactivate = null;
 
@@ -137,7 +138,7 @@ class EditModeTour extends foundry.nue.Tour {
     const AEData = {
       changes: [{
         key: "system.characteristics.Brawn.value",
-        mode: CONST.ACTIVE_EFFECT_MODES.ADD,
+        ...EffectHelpers.changeType(),
         value: 3,
       }],
       name: "example",